requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pymysql==1.1.0

# LLM
groq==0.11.0
google-generativeai>=0.8.0

# Configuration
python-dotenv==1.0.1

# HTTP client
httpx==0.27.2

# Visualization & Frontend
streamlit==1.38.0
plotly==6.1.2
narwhals>=1.15.1
pandas==2.2.0
packaging==24.1

# Development
pytest==8.3.3
black==24.8.0
